+++ girderformindlogger/i18n/locales/en.py
@@ -1,0 +1,12 @@
+locale = {
+    # EMAIL.
+    'invite_email_subject': 'Invitation for an applet',
+    'temporary_access': 'Temporary access',
+
+    # ERRORS.
+    'error_invalid_email': 'Please enter a valid email address',
+    'error_invalid_password': 'Incorrect password for {user} if that user exists',
+
+    # APPLETS.
+    'applet_is_building': "The applet is building. We will send you an email in 10 mins or less when it has been successfully created or failed.",
+}