+++ girderformindlogger/i18n/locales/fr.py
@@ -1,0 +1,12 @@
+locale = {
+    # EMAIL.
+    'invite_email_subject': 'Invitation pour une applet',
+    'temporary_access': 'Accès temporaire',
+
+    # ERRORS.
+    'error_invalid_email': 'Veuillez indiquer une adresse électronique valide',
+    'error_invalid_password': 'Mot de passe incorrect pour {user} si cet utilisateur existe',
+
+    # APPLETS.
+    'applet_is_building': 'L\'applet est en construction. Nous vous enverrons un courriel dans un délai de 10 minutes ou moins lorsqu\'elle aura été créée avec succès ou qu\'elle aura échoué.',
+}