+++ girderformindlogger/models/applet.py
@@ -24,6 +24,7 @@ import json
 import os
 import six
 import threading
+import re
 
 from bson.objectid import ObjectId
 from girderformindlogger import events
@@ -55,7 +56,6 @@ class Applet(FolderModel):
         user=None,
         roles=None,
         constraints=None,
-        appletName=None,
         appletRole='editor',
         accountId=None,
         encryption={}
@@ -87,9 +87,19 @@ class Applet(FolderModel):
             CollectionModel().createCollection('Applets')
             appletsCollection = CollectionModel().findOne({"name": "Applets"})
 
-        appletName = self.validateAppletName(appletName, appletsCollection, accountId)
+        name = self.validateAppletName('%s (0)' % (name), appletsCollection, accountId)
 
         # create new applet
+        metadata = {
+            'protocol': protocol,
+            'applet': constraints if constraints is not None and isinstance(
+                constraints,
+                dict
+            ) else {},
+            'encryption': encryption
+        }
+        metadata['applet']['displayName'] = name
+
         applet = self.setMetadata(
             folder=self.createFolder(
                 parent=appletsCollection,
@@ -98,17 +108,9 @@ class Applet(FolderModel):
                 public=True,
                 creator=user,
                 allowRename=True,
-                appletName=appletName,
                 accountId=accountId
             ),
-            metadata={
-                'protocol': protocol,
-                'applet': constraints if constraints is not None and isinstance(
-                    constraints,
-                    dict
-                ) else {},
-                'encryption': encryption
-            }
+            metadata=metadata
         )
 
         FolderModel().update({
@@ -175,12 +177,7 @@ class Applet(FolderModel):
         )
 
         if 'activities' in formatted:
-            activities = []
-
-            for activity in formatted['activities']:
-                activities.append(ObjectId(formatted['activities'][activity]['_id'].split('/')[-1]))
-
-            self.update({'_id': ObjectId(applet['_id'])}, {'$set': {'meta.protocol.activities': activities}})
+            self.updateActivities(applet, formatted)
 
         # give all roles to creator of an applet
         applet = self.load(applet['_id'], force=True)
@@ -291,14 +288,15 @@ class Applet(FolderModel):
                 return
 
             newInvitation = Invitation().createInvitationForSpecifiedUser(
-                applet,
-                inviter,
-                role,
-                user,
-                user['firstName'],
-                user['lastName'],
-                '',
-                user['email']
+                applet=applet,
+                coordinator=inviter,
+                role=role,
+                user=user,
+                firstName=user['firstName'],
+                lastName=user['lastName'],
+                lang='en',
+                MRN='',
+                userEmail=user['email']
             )
 
             appletProfile = Invitation().acceptInvitation(Invitation().load(newInvitation['_id'], force=True), user, user['email'])
@@ -317,34 +315,44 @@ class Applet(FolderModel):
         from girderformindlogger.models.protocol import Protocol
 
         appletsCollection = CollectionModel().findOne({"name": "Applets"})
-        appletName = self.validateAppletName('{}/'.format(name), appletsCollection, applet['accountId'])
+        appletName = self.validateAppletName(name, appletsCollection, applet['accountId'])
+        prefLabel = appletName
+
+        suffix = re.findall('^(.*?)\s*\((\d+)\)$', appletName)
+        if len(suffix):
+            if suffix[0][0] == applet['meta']['applet']['displayName']:
+                prefLabel = suffix[0][0]
 
         protocolId = applet.get('meta', {}).get('protocol', {}).get('_id', None)
         if not protocolId:
             raise ValidationException('this applet does not have protocol id')
 
-        protocol = Protocol().duplicateProtocol(ObjectId(protocolId.split("/")[1]), editor)
+        protocol = Protocol().duplicateProtocol(ObjectId(protocolId.split("/")[1]), editor, prefLabel)
+
         # create new applet
+        metadata = {
+            **applet.get('meta', {}),
+            'protocol': {
+                '_id': protocol['protocol']['_id'],
+                'activities': [
+                    ObjectId(protocol['activities'][activity]['_id'].split('/')[-1]) for activity in protocol['activities']
+                ],
+                'name': prefLabel
+            }
+        }
+        metadata['applet']['displayName'] = appletName
+
         newApplet = self.setMetadata(
             folder=self.createFolder(
                 parent=appletsCollection,
-                name=name,
+                name=appletName,
                 parentType='collection',
                 public=True,
                 creator=editor,
                 allowRename=True,
-                appletName=appletName,
                 accountId=applet['accountId']
             ),
-            metadata={
-                **applet.get('meta', {}),
-                'protocol': {
-                    '_id': protocol['protocol']['_id'],
-                    'activities': [
-                        ObjectId(protocol['activities'][activity]['_id'].split('/')[-1]) for activity in protocol['activities']
-                    ]
-                }
-            }
+            metadata=metadata
         )
 
         appletGroupName = "Default {} ({})".format(
@@ -528,25 +536,41 @@ class Applet(FolderModel):
 
         return Profile().displayProfileFields(Profile().updateOwnerProfile(applet), thisUser, forceManager=True)
 
-    def validateAppletName(self, appletName, appletsCollection, accountId = None):
-        name = appletName
-        found = False
-        n = 0
-        while found == False:
-            found = True
-            query = {
-                'parentId': appletsCollection['_id'],
-                'appletName': name,
-                'parentCollection': 'collection'
+    def validateAppletName(self, appletName, appletsCollection, accountId = None, currentApplet = None):
+        appletName = appletName.strip()
+
+        suffix = re.findall('^(.*?)\s*\((\d+)\)$', appletName)
+
+        if len(suffix):
+            name = appletName
+            appletName = suffix[0][0]
+            n = int(suffix[0][1])
+        else:
+            n = 0
+
+        if not n:
+            name = appletName
+
+        query = {
+            'parentId': appletsCollection['_id'],
+            'meta.applet.displayName': name,
+            'parentCollection': 'collection'
+        }
+
+        if accountId:
+            query['accountId'] = ObjectId(accountId)
+        if currentApplet and '_id' in currentApplet:
+            query['_id'] = {
+                '$ne': currentApplet['_id']
             }
-            if accountId:
-                query['accountId'] = ObjectId(accountId)
+
+        existing = self.findOne(query)
+        while existing:
+            n = n + 1
+            name = '%s (%d)' % (appletName, n)
+            query['meta.applet.displayName'] = name
 
             existing = self.findOne(query)
-            if existing:
-                found = False
-                n = n + 1
-                name = '%s(%d)' % (appletName, n)
 
         return name
 
@@ -584,19 +608,19 @@ class Applet(FolderModel):
                 user,
                 thread=False,
                 refreshCache=True,
-                meta={'appletId': 'None'}
+                meta={ 'appletId': None },
+                isReloading=False
             )
 
             protocol = protocol[0].get('protocol', protocol[0])
 
-            displayName = Protocol(
-            ).preferredName(
-                protocol
-            )
+            displayName = ''
+            for candidate in ['prefLabel', 'altLabel']:
+                for key in protocol:
+                    if not len(displayName) and key.endswith(candidate) and isinstance(protocol[key], list):
+                        displayName = protocol[key][0]['@value']
 
             name = name if name is not None and len(name) else displayName
-
-            appletName = '{}/'.format(protocolUrl)
 
             applet = self.createApplet(
                 name=name,
@@ -610,12 +634,12 @@ class Applet(FolderModel):
                     ).get(
                         'protocol',
                         {}
-                    ).get('url', protocolUrl)
+                    ).get('url', protocolUrl),
+                    'name': displayName.strip()
                 },
                 user=user,
                 roles=roles,
                 constraints=constraints,
-                appletName=appletName,
                 appletRole=appletRole,
                 accountId=accountId,
                 encryption=encryption
@@ -678,26 +702,25 @@ class Applet(FolderModel):
 
             protocol = protocol.get('protocol', protocol)
 
-            displayName = Protocol(
-            ).preferredName(
-                protocol
-            )
+            displayName = ''
+            for candidate in ['prefLabel', 'altLabel']:
+                for key in protocol:
+                    if not len(displayName) and key.endswith(candidate) and isinstance(protocol[key], list):
+                        displayName = protocol[key][0]['@value']
 
             name = name if name is not None and len(name) else displayName
-
-            appletName = '{}/'.format(protocol.get('@id'))
 
             applet = self.createApplet(
                 name=name,
                 protocol={
                     '_id': 'protocol/{}'.format(
                         str(protocol.get('_id')).split('/')[-1]
-                    )
+                    ),
+                    'name': name.strip()
                 },
                 user=user,
                 roles=roles,
                 constraints=constraints,
-                appletName=appletName,
                 appletRole=appletRole,
                 accountId=accountId,
                 encryption=encryption
@@ -737,6 +760,15 @@ class Applet(FolderModel):
         from girderformindlogger.utility import jsonld_expander
 
         # get a protocol from single json file
+        displayName = protocol['protocol']['data'].get('skos:prefLabel', protocol['protocol']['data'].get('skos:altLabel', '')).strip()
+
+        suffix = re.findall('^(.*?)\s*\((\d+)\)$', displayName)
+        if len(suffix) and applet.get('meta', {}).get('protocol', {}).get('name', '') == suffix[0][0]:
+            protocol['protocol']['data']['skos:prefLabel'] = suffix[0][0]
+        else:
+            applet['meta']['protocol']['name'] = displayName
+            displayName = '%s (0)' % (displayName)
+
         protocol = Protocol().createProtocol(
             protocol,
             user,
@@ -745,9 +777,14 @@ class Applet(FolderModel):
 
         protocol = protocol.get('protocol', protocol)
 
-        displayName = Protocol().preferredName(protocol)
+        applet['meta']['applet']['displayName'] = self.validateAppletName(
+            displayName,
+            CollectionModel().findOne({"name": "Applets"}),
+            accountId,
+            currentApplet = applet
+        )
+        applet = self.setMetadata(folder=applet, metadata=applet['meta'])
 
-        self.save(applet)
         # update appletProfile according to updated applet
         jsonld_expander.clearCache(applet, 'applet')
 
@@ -985,7 +1022,7 @@ class Applet(FolderModel):
 
             if 'url' in metadata['protocol']:
                 metadata['protocol'].pop('url')
-                self.setMetadata(applet, metadata)
+                applet = self.setMetadata(applet, metadata)
 
             jsonld_expander.formatLdObject(protocolFolder, 'protocol', user, refreshCache=True)
         else:
@@ -1056,6 +1093,9 @@ class Applet(FolderModel):
                 "profile": True,
             }))
 
+        if reviewerProfile['_id'] not in reviewerProfile['reviewers'] and (str(reviewerProfile['_id']) in users or not users):
+            profiles.append(reviewerProfile)
+
         query["creatorId"] = {
             "$in": [profile['userId'] for profile in profiles]
         }
@@ -1077,7 +1117,7 @@ class Applet(FolderModel):
         IRIs = {}
         # IRIs refers to available versions for specified IRI
         # IRI is github url for items created by url, and pair of activity id and item id for items created by applet-builder
-        # ex: IRIS = { 
+        # ex: IRIS = {
         #                'https://raw.githubusercontent.com/ChildMindInstitute/TokenLogger_applet/master/activities/TokenActivity/items/token_screen': ['0.0.1'],
         #                 '5f87e250c3942f7d5df7b7ca/5f87e25ac3942f7d5df7b7ce': ['0.0.2', '0.0.3']
         #            }
@@ -1119,7 +1159,7 @@ class Applet(FolderModel):
 
         data.update(
             Protocol().getHistoryDataFromItemIRIs(
-                applet.get('meta', {}).get('protocol', {}).get('_id', '').split('/')[-1], 
+                applet.get('meta', {}).get('protocol', {}).get('_id', '').split('/')[-1],
                 IRIs
             )
         )
@@ -1246,12 +1286,21 @@ class Applet(FolderModel):
         if protocolUrl is None:
             raise AccessException('this applet is not uploaded from url')
 
+        protocol = Protocol().findOne({
+            '_id': ObjectId(applet.get('meta', {}).get('protocol', {}).get('_id' , '').split('/')[-1])
+        })
+
+        if 'appletId' not in protocol.get('meta', {}):
+            protocol['meta']['appletId'] = 'None'
+            Protocol().setMetadata(protocol, protocol['meta'])
+
         protocol = Protocol().getFromUrl(
             protocolUrl,
             'protocol',
             editor,
             thread=False,
-            refreshCache=True
+            refreshCache=True,
+            meta={'appletId': protocol['meta']['appletId']},
         )
 
         protocol = protocol[0].get('protocol', protocol[0])
@@ -1260,16 +1309,41 @@ class Applet(FolderModel):
             if 'meta' in applet and 'protocol' in applet['meta']:
                 applet['meta']['protocol']['_id'] = protocol['_id']
 
+            displayName = ''
+            for candidate in ['prefLabel', 'altLabel']:
+                for key in protocol:
+                    if not len(displayName) and key.endswith(candidate) and isinstance(protocol[key], list):
+                        displayName = protocol[key][0]['@value']
+
+            #suffix = re.findall('^(.*?)\s*\((\d+)\)$', applet.get('meta', {}).get('applet', {}).get('displayName', {}))
+            #if len(suffix):
+            #    displayName = '%s (%s)' % (displayName, suffix[0][1])
+
+            applet['meta']['applet']['displayName'] = self.validateAppletName(
+                applet['displayName'],
+                CollectionModel().findOne({"name": "Applets"}),
+                accountId = applet['accountId'],
+                currentApplet = applet
+            )
+
+        self.save(applet)
+
         from girderformindlogger.utility import jsonld_expander
 
         jsonld_expander.clearCache(applet, 'applet')
-        jsonld_expander.formatLdObject(
+        print('Cache clear')
+
+        formatted = jsonld_expander.formatLdObject(
             applet,
             'applet',
             editor,
             refreshCache=False,
             responseDates=False
         )
+
+        if 'activities' in formatted:
+            activities = self.updateActivities(applet, formatted)
+            Profile().update_profile_activities_by_applet_id(applet, activities)
 
     def getAppletsForUser(self, role, user, active=True, idOnly = False):
         """
@@ -1517,3 +1591,12 @@ class Applet(FolderModel):
                 raise ValidationException(
                     "Invalid Applet ID."
                 )
+
+    def updateActivities(self, applet, obj):
+        activities = [ObjectId(obj['activities'][activity]['_id'].split('/')[-1])
+                      for activity in obj.get('activities', [])]
+
+        self.update({'_id': ObjectId(applet['_id'])},
+                    {'$set': {'meta.protocol.activities': activities}})
+
+        return activities